댓글 추가, 삭제 후 리렌더링 되도록 함

+++ src/api/commentApi.ts
@@ -26,11 +26,11 @@ export const enterComment = async (text: string, video_id: string) => {
 }
 
 // 댓글 삭제 API
-export const deleteComment = async (commentId: number) => {
+export const deleteComment = async (commentId: string) => {
   const { data, error } = await supabaseAdmin
     .from("video_comment")
     .delete()
-    .eq("id", commentId)
+    .eq("anonymous_user_id", commentId)
 
   if (error) {
     console.error("Error deleting comment:", error.message)
@@ -49,7 +49,7 @@ export const readComment = async () => {
     if (error) {
       console.error(`데이터 통신에 실패하였습니다..😵‍💫 ${error.message}`)
     } else {
-      console.log("Supabase 데이터 가져오기 성공:", data)
+      // console.log("Supabase 데이터 가져오기 성공:", data)
       return data
     }
   } catch (error) {
